requirements.txt
discord.py
aiosqlite
python-dotenv
yfinance
matplotlib
pandas
mplfinance
aiofiles
aiohttp
tweepy

